Reduce complexity and improve readability and maintainability of addrun code
